framework for ipc
